+++ lib/src/features/main_menu/data/datasource/remote/main_menu_datasource_abst.dart
@@ -1,6 +1,6 @@
 import '../../../../../shared/data/entities/user.dart';
-import '../../../domain/models/profile_user_params.dart';
-import '../../../domain/models/security_contacts_params.dart';
+import '../../../domain/models/profile/profile_user_params.dart';
+import '../../../domain/models/contacts/security_contacts_params.dart';
 import '../../entities/security_contacts.dart';
 
 abstract class MainMenuRemoteDatasource {
