+++ lib/src/features/main_menu/data/datasource/remote/main_menu_datasource_impl.dart
@@ -5,10 +5,10 @@ import '../../../../../core/network/error/dio_error_handler.dart';
 import '../../../../../core/network/error/exceptions.dart';
 import '../../../../../core/network/network_url.dart';
 import '../../../../../shared/data/entities/user.dart';
-import '../../../domain/models/profile_user_model.dart';
-import '../../../domain/models/profile_user_params.dart';
-import '../../../domain/models/security_contacts_model.dart';
-import '../../../domain/models/security_contacts_params.dart';
+import '../../../domain/models/profile/profile_user_model.dart';
+import '../../../domain/models/profile/profile_user_params.dart';
+import '../../../domain/models/contacts/security_contacts_model.dart';
+import '../../../domain/models/contacts/security_contacts_params.dart';
 import '../../entities/security_contacts.dart';
 import 'main_menu_datasource_abst.dart';
