+++ lib/src/features/main_menu/data/repository/main_menu_repository_impl.dart
@@ -4,8 +4,8 @@ import 'package:dartz/dartz.dart';
 import '../../../../core/network/error/exceptions.dart';
 import '../../../../core/network/error/failures.dart';
 import '../../../../shared/data/entities/user.dart';
-import '../../domain/models/profile_user_params.dart';
-import '../../domain/models/security_contacts_params.dart';
+import '../../domain/models/profile/profile_user_params.dart';
+import '../../domain/models/contacts/security_contacts_params.dart';
 import '../../domain/repository/main_menu_repository_abst.dart';
 import '../datasource/remote/main_menu_datasource_abst.dart';
 import '../entities/security_contacts.dart';
