+++ lib/src/features/main_menu/domain/models/profile/profile_user_model.dart
@@ -1,4 +1,4 @@
-import '../../../../shared/data/entities/user.dart';
+import '../../../../../shared/data/entities/user.dart';
 
 class UserModel extends User {
   UserModel(
