+++ lib/src/features/main_menu/domain/repository/main_menu_repository_abst.dart
@@ -3,8 +3,8 @@ import 'package:dartz/dartz.dart';
 import '../../../../core/network/error/failures.dart';
 import '../../../../shared/data/entities/user.dart';
 import '../../data/entities/security_contacts.dart';
-import '../models/profile_user_params.dart';
-import '../models/security_contacts_params.dart';
+import '../models/profile/profile_user_params.dart';
+import '../models/contacts/security_contacts_params.dart';
 
 abstract class MainMenuRepository {
   Future<Either<Failure, List<SecurityContacts>>> getSecurityContacts(
