+++ lib/src/features/main_menu/domain/use_cases/contacts/get_contacts_usecase.dart
@@ -2,9 +2,9 @@ import 'package:dartz/dartz.dart';
 import 'package:drive_on/src/core/network/error/failures.dart';
 import 'package:drive_on/src/core/utils/usecases/usecase.dart';
 import 'package:drive_on/src/features/main_menu/data/entities/security_contacts.dart';
-import 'package:drive_on/src/features/main_menu/domain/models/security_contacts_params.dart';
+import 'package:drive_on/src/features/main_menu/domain/models/contacts/security_contacts_params.dart';
 
-import '../repository/main_menu_repository_abst.dart';
+import '../../repository/main_menu_repository_abst.dart';
 
 class GetContactsUseCase
     extends UseCase<List<SecurityContacts>, SecurityContactsParams> {
