+++ lib/src/features/main_menu/domain/use_cases/profile/get_profile_use_case.dart
@@ -1,10 +1,10 @@
 import 'package:dartz/dartz.dart';
 
-import '../../../../core/network/error/failures.dart';
-import '../../../../core/utils/usecases/usecase.dart';
-import '../../../../shared/data/entities/user.dart';
-import '../models/profile_user_params.dart';
-import '../repository/main_menu_repository_abst.dart';
+import '../../../../../core/network/error/failures.dart';
+import '../../../../../core/utils/usecases/usecase.dart';
+import '../../../../../shared/data/entities/user.dart';
+import '../../models/profile/profile_user_params.dart';
+import '../../repository/main_menu_repository_abst.dart';
 
 class GetProfileUseCase extends UseCase<User, ProfileUserParams> {
   final MainMenuRepository repository;
