+++ lib/src/features/main_menu/presentation/cubit/contacts_cubit/contacts_cubit.dart
@@ -1,8 +1,8 @@
-import 'package:drive_on/src/features/main_menu/domain/use_cases/get_contacts_usecase.dart';
+import 'package:drive_on/src/features/main_menu/domain/use_cases/contacts/get_contacts_usecase.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../../../../core/utils/constants/error_constants.dart';
-import '../../../domain/models/security_contacts_params.dart';
+import '../../../domain/models/contacts/security_contacts_params.dart';
 import 'contacts_state.dart';
 
 class ContactsCubit extends Cubit<ContactsState> {
