+++ lib/src/features/main_menu/presentation/pages/main_menu_pag.dart
@@ -12,8 +12,8 @@ import '../../../../shared/presentation/widgets/custom_progress_indicator.dart';
 import '../../../../shared/presentation/widgets/dynamic_bottom_bar.dart';
 import '../../../../shared/presentation/widgets/floating_snack_bars.dart';
 import '../../data/datasource/remote/main_menu_datasource_impl.dart';
-import '../../domain/use_cases/get_contacts_usecase.dart';
-import '../../domain/use_cases/get_profile_use_case.dart';
+import '../../domain/use_cases/contacts/get_contacts_usecase.dart';
+import '../../domain/use_cases/profile/get_profile_use_case.dart';
 import '../cubit/contacts_cubit/contacts_cubit.dart';
 import '../cubit/contacts_cubit/contacts_state.dart';
 import '../cubit/profile_cubit/profile_cubit.dart';
